# admin/docs/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
...
  baseModel.DocsAuthLg:
    properties:
      code:
        description: 响应码
        example: E002
        type: string
      msg:
        description: 响应消息
        example: Not currently logged in
        type: string
    type: object
  baseModel.DocsAuthNg:
    properties:
      code:
        description: 响应码
        example: E003
        type: string
      msg:
        description: 响应消息
        example: Access to the current interface is prohibited
        type: string
    type: object
  baseModel.DocsErr:
    properties:
      code:
        description: 响应码
        example: E000
        type: string
      msg:
        description: 响应消息
        example: System exception
        type: string
    type: object
  baseModel.DocsOk:
    properties:
      code:
        description: 响应码
        example: SX0X/FX0X
        type: string
      data:
        description: 响应数据
        example: Response data
        type: string
      msg:
        description: 响应消息
        example: Success/Fail
        type: string
      unPop:
        description: 免弹窗提示
        example: true
        type: boolean
    type: object
  baseModel.DocsReq:
    properties:
      httpCode:
        description: 响应HTTPCode，不传默认响应200，支持200/400/401/403/500
        example: 200
        type: integer
    type: object
  baseModel.DocsVail:
    properties:
      code:
        description: 响应码
        example: E001
        type: string
      msg:
        description: 响应消息
        example: xx Field should be required
        type: string
    type: object
  baseModel.IdReq:
    properties:
      id:
        description: 数据ID
        example: 1
        type: integer
    type: object
...
  baseModel.Req:
    type: object
  baseModel.ResBody:
    properties:
      code:
        description: 响应码
        example: S000/F000
        type: string
      data:
        description: 响应数据
      msg:
        description: 响应消息
        example: Success/Fail
        type: string
      unPop:
        description: 免弹窗提示
        example: true
        type: boolean
    type: object
...
  baseModel.SortReq:
    properties:
      id:
        description: 数据ID
        example: 1
        type: integer
      sort:
        description: 序号
        example: 1
        type: integer
    required:
    - id
    type: object
  baseModel.SortRes:
    properties:
      id:
        description: 默认数据ID
        example: 1
        type: integer
      name:
        description: 权限名称，界面展示，建议与界面导航一致
        example: Account
        type: string
      sort:
        description: 权限排序
        example: 1
        type: integer
    type: object
  baseModel.Tree:
    properties:
      children:
        description: 子树
        items:
          $ref: '#/definitions/baseModel.Tree'
        type: array
      expand:
        description: 拓展信息
      id:
        description: 表示树数据ID
        example: 1
        type: integer
      key:
        description: 树键
        example: ROOT
        type: string
      level:
        description: 表示树等级
        example: "1"
        type: string
      title:
        description: 树标题
        example: ROOT
        type: string
    type: object
...
  platModel.DeptAccountRes:
    properties:
      account:
        description: 账号
        example: demo
        type: string
      id:
        description: 数据ID
        example: 1
        type: integer
      name:
        description: 姓名
        example: demo
        type: string
    type: object
  platModel.DeptAddReq:
    properties:
      name:
        description: 部门名称
        example: demo
        maxLength: 16
        type: string
      permissionType:
        description: 权限类型，枚举：0_本部门与子部门 1_本部门 2_个人 3_全局 4_指定部门 5_指定人
        enum:
        - "0"
        - "1"
        - "2"
        - "3"
        - "4"
        - "5"
        example: "0"
        type: string
      pid:
        description: 父级部门ID，租户创建时默认创建根部门，父级ID=0
        example: 0
        type: integer
    required:
    - name
    - permissionType
    - pid
    type: object
  platModel.DeptEditReq:
    properties:
      id:
        description: 数据ID
        example: 1
        type: integer
      name:
        description: 部门名称
        example: demo
        maxLength: 16
        type: string
      permissionType:
        description: 权限类型，枚举：0_本部门与子部门 1_本部门 2_个人 3_全局 4_指定部门 5_指定人
        enum:
        - "0"
        - "1"
        - "2"
        - "3"
        - "4"
        - "5"
        example: "0"
        type: string
      pid:
        description: 父级部门ID，租户创建时默认创建根部门，父级ID=0
        example: 0
        type: integer
    required:
    - id
    - name
    - permissionType
    - pid
    type: object
  platModel.DeptGetRes:
    properties:
      accounts:
        description: 部门账号，第1组是领导，第二组是成员
        items:
          items:
            $ref: '#/definitions/platModel.DeptAccountRes'
          type: array
        type: array
      id:
        description: 数据ID
        example: 1
        type: integer
      mark:
        description: 变更标识，枚举：0_可变更 1_禁止变更
        example: "0"
        type: string
      name:
        description: 部门名称
        example: demo
        type: string
      permissionType:
        description: 权限类型，枚举：0_本部门与子部门 1_本部门 2_个人 3_指定部门 4_指定人 5_全局
        example: "0"
        type: string
      pid:
        description: 父级部门ID，租户创建时默认创建根部门，父级ID=0
        example: 0
        type: integer
      sort:
        description: 同级部门排序
        example: 0
        type: integer
    type: object
  platModel.DeptToReq:
    properties:
      id:
        description: 当前部门数据ID
        example: 1
        type: integer
      toId:
        description: 目标部门数据ID
        example: 1
        type: integer
      toType:
        description: 迁移模式，枚举：0_并入（子部门形式） 1_移交（部门保留，成员和子部门移交给新部门）
        example: "0"
        type: string
    required:
    - id
    - toId
    - toType
    type: object
...
host: localhost:8000
info:
  contact:
    email: dev@siteol.com
    name: smart
    url: https://smart.siteol.com
  description: "![Smart](http://localhost:8000/docs/file/logo.png)\r\n\r\n# 概述\r\n\r\n轻量易用的授权基座smart，提供一个非官方RBAC授权模型的基础开箱即用中台管理服务。\r\n\r\n支持响应码自定义国际化、高粒度权限分配、接口路由配置和实时生效，同时支持部门数据权限横向访问限制。\r\n\r\n本系统的全部接口采用【POST】【application/json】方式传输数据。\r\n\r\n除开放接口以外的其他接口均需要通过【ApiKeyAuth:请求头[Token]】完成鉴权。\r\n\r\n系统技术栈：Golang、VueNext、MySQL、Redis、Gin、ArcoDesign\r\n\r\n#
    接口文档\r\n\r\n轻度依赖集成，通过静态HTML加载yaml文件进行打开接口文档。\r\n\r\n该接口文档提供Swagger[支持调试]和ReDoc[阅读增强]两个版本。\r\n\r\n[Swagger[支持调试]：/docs/swagger/index.html](http://localhost:8000/docs/swagger/index.html)
    丨 [ReDoc[阅读增强]：/docs/redoc/index.html](http://localhost:8000/docs/redoc/index.html)\r\n\r\n##
    生成说明\r\n\r\n- 首次集成\r\n\r\ngo install github.com/swaggo/swag/cmd/swag@latest\r\n\r\n-
    格式化注释代码 (选用)\r\n\r\nswag fmt\r\n\r\n- 初始化swagger.yaml文件\r\n\r\nswag init\r\n\r\n-
    如果像本工程一样依赖了api.md和独立的api.go文件描述项目（推荐，后续持续使用）\r\n\r\nswag init -g api.go --md .\r\n\r\n-
    删除多余的生成\r\n\r\nrm .\\docs\\docs.go\r\n\r\n删除docs.go是因为本项目中并未采用下述依赖来集成。\r\n\r\n-
    github.com/swaggo/swag\r\n- github.com/swaggo/gin-swagger\r\n- github.com/swaggo/files\r\n\r\n而是通过HTML模板+JS+YAML引入集成，移除docs.go用于避免依赖编译报错。\r\n\r\n源帮助页：[https://github.com/swaggo/swag/blob/master/README_zh-CN.md](https://github.com/swaggo/swag/blob/master/README_zh-CN.md)\r\n"
  title: smart
  version: "1.0"
  x-logo:
    altText: smart
...
paths:
...
  /docs/sample:
    post:
      consumes:
      - application/json
      description: 系统API基本示例
      operationId: Sample示例
      parameters:
      - description: 示例请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/baseModel.DocsReq'
      produces:
      - application/json
      responses:
        "200":
          description: 业务受理成功
          schema:
            $ref: '#/definitions/baseModel.DocsOk'
        "400":
          description: 数据校验失败
          schema:
            $ref: '#/definitions/baseModel.DocsVail'
        "401":
          description: 当前尚未登陆
          schema:
            $ref: '#/definitions/baseModel.DocsAuthLg'
        "403":
          description: 权限校验失败
          schema:
            $ref: '#/definitions/baseModel.DocsAuthNg'
        "500":
          description: 服务系统异常
          schema:
            $ref: '#/definitions/baseModel.DocsErr'
      security:
      - Token: []
      summary: 通用API示例
      tags:
      - 开放接口
...
  /plat/dept/add:
    post:
      consumes:
      - application/json
      description: 新建集团部门
      operationId: AddDept 集团部门新建
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/platModel.DeptAddReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - Token: []
      summary: 集团部门新建
      tags:
      - 集团部门
  /plat/dept/bro:
    post:
      consumes:
      - application/json
      description: 同级部门列表
      operationId: BroDept 同级部门
      parameters:
      - description: 语言，不传默认为zh-CN
        in: header
        name: Lang
        type: string
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/baseModel.IdReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/baseModel.SortRes'
                  type: array
              type: object
      security:
      - Token: []
      summary: 同级部门
      tags:
      - 集团部门
  /plat/dept/del:
    post:
      consumes:
      - application/json
      description: 集团部门移除处理
      operationId: DelDept 集团部门移除
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/baseModel.IdReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - Token: []
      summary: 集团部门移除
      tags:
      - 集团部门
  /plat/dept/edit:
    post:
      consumes:
      - application/json
      description: 基于数据ID编辑集团部门
      operationId: EditDept 集团部门编辑
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/platModel.DeptEditReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - Token: []
      summary: 集团部门编辑
      tags:
      - 集团部门
  /plat/dept/get:
    post:
      consumes:
      - application/json
      description: 查询集团部门详情
      operationId: GetDept 集团部门详情
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/baseModel.IdReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  $ref: '#/definitions/platModel.DeptGetRes'
              type: object
      security:
      - Token: []
      summary: 集团部门详情
      tags:
      - 集团部门
  /plat/dept/sort:
    post:
      consumes:
      - application/json
      description: 同级部门排序功能
      operationId: SortDept部门排序
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          items:
            $ref: '#/definitions/baseModel.SortReq'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - Token: []
      summary: 部门排序
      tags:
      - 集团部门
  /plat/dept/to:
    post:
      consumes:
      - application/json
      description: 部门以及子部门迁移到新部门
      operationId: ToDept 部门迁移
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/platModel.DeptToReq'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - Token: []
      summary: 部门迁移
      tags:
      - 集团部门
  /plat/dept/tree:
    post:
      consumes:
      - application/json
      description: 查询集团部门树数据
      operationId: TreeDept 集团部门树
      parameters:
      - description: 请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/baseModel.Req'
      produces:
      - application/json
      responses:
        "200":
          description: 响应成功
          schema:
            allOf:
            - $ref: '#/definitions/baseModel.ResBody'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/baseModel.Tree'
                  type: array
              type: object
      security:
      - Token: []
      summary: 集团部门树
      tags:
      - 访问部门
...
securityDefinitions:
  Token:
    in: header
    name: Token
    type: apiKey
swagger: "2.0"
...